Display avatar left of comments by default
Previously avatars would be displayed on top of comments. Most comment
and forum systems will display avatars to the left of comments/posts.
The https://github.com/snorremd/talk-plugin-gravatar plugin currently
renders avatars in the avatar slot which will render the avatar on top
of the comments. This change modified the comment container to be
displayed as a flex container with row direction. It also adds 10px top
margin to the avatar slot which corresponds with the comment header.

Signed-off-by: Snorre Magnus Davøen <snorremd@gmail.com>