Store cursor state in relay
- currentCursor depicts the current "NEW" dividing cursor
we see in our live chat until we reload or jump around the
stream
- latestCursor stores the latest cursor we have actively
read to in the stream. This allows us to always access both
the current new marker we're showing as well as how far the
viewer has traversed in the stream past that point.

The latestCursor is persisted to both the local relay but
also the localStorage, so when the stream is reloaded, that
will become the new currentCursor.

Similarly, when the user jumps around in the stream via a
refresh or jump to live, the currentCursor is set along with
the pagination state set in the LiveTabQuery.