Add options for scrolling to footer when setting cursor
This allows us to conditionally scroll the chat footer into
view depending on where we're jumping in the stream.

Prior to the "go to start" functionality, always scrolling the
footer into view after setting the cursor was correct, but
when jumping to the start of the stream, we don't want to do this
and doing so can cause use to scroll away from the iframe if
the scroll bubbles up into the main page.

This allows us to do both, by default, setting a cursor will make
sure the footer is visible, but for other stream operations we
can conditionally tell it not to do so.