Switch from url-regex to url-regex-safe
Use url-regex-safe to fix ReDoS warning https://nvd.nist.gov/vuln/detail/CVE-2020-7661. However, this does not actually improve security, since URL_REGEX is only used on the client, not the server, and url-regex-safe uses the linear-time RE2 only in nodejs and still uses the built-in exponential-time RegExp on the browser.