Fix issue where archive job could end abruptly
Previous version of batch iteration was a .forEach
whereas now it is a for of loop. We can't return out
of a for..of without breaking the batch loop.

Instead switch to checking on lock prior to archiving
so we don't need a return/continue and add comments
to clarify how this works.