Update image to node:14 LTS and disable sudo npm update -g
We can't be using sudo to do npm updates as it modifies
the ~/.npm and node_modules/ ownership. Instead we need to
use the CircleCI images as-is and regularly update to the
latest LTS as they come out.

This will give us the security of a stable and secure npm
command as well as prevent permissions issues around the
dependencies breaking our builds and causing `npm ci` to
hang.