Stop converting html to text for mail content
Converting the html rendered emails to the markup
like syntax that email can cause the email body
to be parsed incorrectly when sent. This is because
sometimes we embed a JWT in a link on our emails and
the email text body syntax has more special character
restrictions than HTML.

Sometimes, when a JWT is encoded to create a password
reset token URL, it uses some of these invalid
characters. This makes the mail parser abort the render
of the message and delete anything after the invalid
character.

If we use html in the text body for the email, it will
render correctly as the mail processor will recognize
that the body is html and leave the message as is and
render as html instead of using mail markup. This ensures
the reset URL always works and is handled correctly.

Almost every mail client knows how to process html
when rendering the content body so this should be
fine.