SSO: always ban bad domain users, auto-premod only on newly created user
This splits the checks for auto premod/ban for SSO/local auth users.
Auto-banning will always be checked for SSO users, so that when a bad
domain email is found, it is always auto-banned.

Pre-moderation is only set on the creation of a new SSO account. This allows
the pre-moderation to automatically be set, but if a mod/admin decides to remove
the pre-moderation later, they can and the system won't automatically re-premod
the user because it was in the domain list.