Add archiveComments index for downloading comments
There were several slow queries happening when a user was
downloading their comments and some of their comments are
archived. The reason being that archivedComments did not
have the same indexes as the regular comments for making
this lookup speedy.

Jessie has added the index to the main archiving prod
clusters, but for future reference, the index has been
added to the INDEXES.md.