Add extra time for `load more comments` test to rune
This test can fail if the test runner is bogged down running other
tests cause it doesn't have enough resources to render the whole
stream + lots of extra comments. If we give it a bit of extra time
though, it always passes.

It also always passes within the allotted time if it's the only test
running, so it's not like the test is bad, just timing out.