Make jobStatus handler POST instead of GET so it works with HTTPS
Despite this working locally even when runnin in Docker, if you
use HTTPS with HTML/2 protocol, it will reject the GET with a body
calling it a "bad frame" error on the request. So we use post so
it'll stop whining and blocking our attempts to retrieve the job
progress.