Fix validation on story tree generation jobs so they run with optional/empty inputs (#3996)
* Make jobStatus handler POST instead of GET so it works with HTTPS

Despite this working locally even when runnin in Docker, if you
use HTTPS with HTML/2 protocol, it will reject the GET with a body
calling it a "bad frame" error on the request. So we use post so
it'll stop whining and blocking our attempts to retrieve the job
progress.

* Up jobStatus rate limiting to 10 times a minute

This is just hitting redis now, so it's okay to let it
take a lot more traffic. Though putting it to 1 request
every 10 seconds should be enough for even the spammiest
of progress grabbers.

* Allow empy strings in disableCommentingMessage

* Validate regenerate story trees job data before accepting job

This allows us to return to the user whether their input is bad
before we start the job so they can know to re-submit if it fails
to validate.