Expose the tag count types so we can consolidate tag checks
The resolvers were using a subset of the total GQLTag types
that are tracked in the story tags count tally. By exposing
the remaining tags, we don't have to check for both the subset
and the superset. We can simply use the GQL generated tag
counts everywhere. This simplifies the typing for the tag
counts.