add dependencies so docker builds can run on M1 macs
Alpine linux relies on host OS having the right binaries
for it to build our node app. Since it sees the pre-build
x86 libs, it panics and tries to find/build them locally.

They aren't there cause none of this pre-built for Apple M1.
By installing them, we ensure they are available for building
our apps dependencies no matter which platform we are on.

Please note, this is only for our dependencies to build the
app, this does not affect our runtime dependencies such as re2
for our wordlist detection which is pulled in as an x86 lib
for use on our cloud systems which are x86 based.