update story url if valid id is provided
During the retrieval of a comment stream via the
createStreamEmbed, if the site provides a valid
storyID that matches a real story but the url it
is coming from differs from the one on record, we
update the story url.

This prevents numerous "DUPLICATE_STORY_ID" errors
when newsrooms modify their story URL multiple times
in their CMS due to article title changes,
slug changes, etc.